requirements.txt
colorama==0.4.6
comtypes==1.4.6
concurrent_log_handler==0.9.25
pywin32==305
PyYAML==6.0.2
tabulate==0.9.0
